requirements.txt
pillow
customtkinter
pyperclip
googletrans
pytesseract